backend/requirements.txt
celery
python-docx
Flask
Flask_Cors
jieba
PyJWT
Werkzeug
eventlet

psutil
requests

# 测试依赖
pytest
pytest-flask
pytest-mock
pytest-cov
requests-mock
